Extract docker package from get-kind-cidr-prefix.go.
